+++ packages/replace/src/index.js
@@ -24,6 +24,7 @@ function getReplacements(options) {
   delete values.exclude;
   delete values.sourcemap;
   delete values.sourceMap;
+  delete values.objectGuards;
   return values;
 }
 
@@ -35,19 +36,61 @@ function mapToFunctions(object) {
   }, {});
 }
 
+const objKeyRegEx =
+  /^([_$a-zA-Z\xA0-\uFFFF][_$a-zA-Z0-9\xA0-\uFFFF]*)(\.([_$a-zA-Z\xA0-\uFFFF][_$a-zA-Z0-9\xA0-\uFFFF]*))+$/;
+function expandTypeofReplacements(replacements) {
+  Object.keys(replacements).forEach((key) => {
+    const objMatch = key.match(objKeyRegEx);
+    if (!objMatch) return;
+    let dotIndex = objMatch[1].length;
+    let lastIndex = 0;
+    do {
+      // eslint-disable-next-line no-param-reassign
+      replacements[`typeof ${key.slice(lastIndex, dotIndex)} ===`] = '"object" ===';
+      // eslint-disable-next-line no-param-reassign
+      replacements[`typeof ${key.slice(lastIndex, dotIndex)} !==`] = '"object" !==';
+      // eslint-disable-next-line no-param-reassign
+      replacements[`typeof ${key.slice(lastIndex, dotIndex)}===`] = '"object"===';
+      // eslint-disable-next-line no-param-reassign
+      replacements[`typeof ${key.slice(lastIndex, dotIndex)}!==`] = '"object"!==';
+      // eslint-disable-next-line no-param-reassign
+      replacements[`typeof ${key.slice(lastIndex, dotIndex)} ==`] = '"object" ===';
+      // eslint-disable-next-line no-param-reassign
+      replacements[`typeof ${key.slice(lastIndex, dotIndex)} !=`] = '"object" !==';
+      // eslint-disable-next-line no-param-reassign
+      replacements[`typeof ${key.slice(lastIndex, dotIndex)}==`] = '"object"===';
+      // eslint-disable-next-line no-param-reassign
+      replacements[`typeof ${key.slice(lastIndex, dotIndex)}!=`] = '"object"!==';
+      lastIndex = dotIndex + 1;
+      dotIndex = key.indexOf('.', lastIndex);
+    } while (dotIndex !== -1);
+  });
+}
+
 export default function replace(options = {}) {
   const filter = createFilter(options.include, options.exclude);
-  const { delimiters } = options;
-  const functionValues = mapToFunctions(getReplacements(options));
-  const keys = Object.keys(functionValues)
-    .sort(longest)
-    .map(escape);
-  const pattern = delimiters
-    ? new RegExp(`${escape(delimiters[0])}(${keys.join('|')})${escape(delimiters[1])}`, 'g')
-    : new RegExp(`\\b(${keys.join('|')})\\b`, 'g');
+  const { delimiters = ['\\b', '\\b(?!\\.)'], preventAssignment, objectGuards } = options;
+  const replacements = getReplacements(options);
+  if (objectGuards) expandTypeofReplacements(replacements);
+  const functionValues = mapToFunctions(replacements);
+  const keys = Object.keys(functionValues).sort(longest).map(escape);
+  const lookahead = preventAssignment ? '(?!\\s*=[^=])' : '';
+  const pattern = new RegExp(
+    `${delimiters[0]}(${keys.join('|')})${delimiters[1]}${lookahead}`,
+    'g'
+  );
 
   return {
     name: 'replace',
+
+    buildStart() {
+      if (![true, false].includes(preventAssignment)) {
+        this.warn({
+          message:
+            "@rollup/plugin-replace: 'preventAssignment' currently defaults to false. It is recommended to set this option to `true`, as the next major version will default this option to `true`."
+        });
+      }
+    },
 
     renderChunk(code, chunk) {
       const id = chunk.fileName;