+++ packages/replace/test/sourcemaps.js
@@ -2,10 +2,15 @@
 
 const test = require('ava');
 const { rollup } = require('rollup');
-const { SourceMapConsumer } = require('source-map');
 const { getLocator } = require('locate-character');
+// source-map uses the presence of fetch to detect browser environments which
+// breaks in Node 18
+const { fetch } = global;
+delete global.fetch;
+const { SourceMapConsumer } = require('source-map');
 
-const replace = require('../dist/rollup-plugin-replace.cjs.js');
+global.fetch = fetch;
+const replace = require('..');
 
 const { getOutputFromGenerated } = require('./helpers/util');
 
@@ -44,7 +49,7 @@ test('generates sourcemaps by default', async (t) => {
     onwarn(warning) {
       throw new Error(warning.message);
     },
-    plugins: [replace({ values: { ANSWER: '42' } }), testPlugin]
+    plugins: [replace({ values: { ANSWER: '42' }, preventAssignment: true }), testPlugin]
   });
 
   const { code, map } = getOutputFromGenerated(
@@ -60,7 +65,14 @@ test('generates sourcemaps if enabled in plugin', async (t) => {
     onwarn(warning) {
       throw new Error(warning.message);
     },
-    plugins: [replace({ values: { ANSWER: '42' }, sourcemap: true }), testPlugin]
+    plugins: [
+      replace({
+        values: { ANSWER: '42' },
+        sourcemap: true,
+        preventAssignment: true
+      }),
+      testPlugin
+    ]
   });
 
   const { code, map } = getOutputFromGenerated(
@@ -76,7 +88,14 @@ test('generates sourcemaps if enabled in plugin (camelcase)', async (t) => {
     onwarn(warning) {
       throw new Error(warning.message);
     },
-    plugins: [replace({ values: { ANSWER: '42' }, sourceMap: true }), testPlugin]
+    plugins: [
+      replace({
+        values: { ANSWER: '42' },
+        sourceMap: true,
+        preventAssignment: true
+      }),
+      testPlugin
+    ]
   });
 
   const { code, map } = getOutputFromGenerated(
@@ -98,7 +117,14 @@ test('does not generate sourcemaps if disabled in plugin', async (t) => {
       );
       warned = true;
     },
-    plugins: [replace({ values: { ANSWER: '42' }, sourcemap: false }), testPlugin]
+    plugins: [
+      replace({
+        values: { ANSWER: '42' },
+        sourcemap: false,
+        preventAssignment: true
+      }),
+      testPlugin
+    ]
   });
 
   t.truthy(!warned);
@@ -118,7 +144,14 @@ test('does not generate sourcemaps if disabled in plugin (camelcase)', async (t)
       );
       warned = true;
     },
-    plugins: [replace({ values: { ANSWER: '42' }, sourceMap: false }), testPlugin]
+    plugins: [
+      replace({
+        values: { ANSWER: '42' },
+        sourceMap: false,
+        preventAssignment: true
+      }),
+      testPlugin
+    ]
   });
 
   t.truthy(!warned);